+++ src/components/Navigation/Pages/PageHarvest/index.js
@@ -1,0 +1,1 @@
+export { default } from './PageHarvest'
