+++ src/components/RehydrationProvider.js
@@ -21,7 +21,7 @@ import history from '../config/history'
 import { configureStore } from '../configureStore'
 import { Provider } from 'react-redux'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import { App, PageHarvest } from './'
+import { App } from './'
 import { omit } from 'lodash'
 import PageAbout from './PageAbout'
 import PageContribution from './Navigation/Pages/PageContribution'
@@ -32,6 +32,7 @@ import PageBrowse from './Navigation/Pages/PageBrowse'
 import PageStats from './Navigation/Pages/PageStats'
 import PageStatus from './Navigation/Pages/PageStatus'
 import PageFile from './Navigation/Pages/PageFile'
+import PageHarvest from './Navigation/Pages/PageHarvest'
 
 const store = configureStore()
