+++ src/components/index.js
@@ -34,7 +34,6 @@ export { default as PyPiVersionPicker } from './PyPiVersionPicker'
 export { default as RubyGemsSelector } from './RubyGemsSelector'
 export { default as RubyGemsVersionPicker } from './RubyGemsVersionPicker'
 export { default as PageAbout } from './PageAbout'
-export { default as PageHarvest } from './PageHarvest'
 export { default as RehydrationProvider } from './RehydrationProvider'
 export { default as RowEntityList } from './RowEntityList'
 export { default as Section } from './Section'
